LayoutTests/crypto/subtle/aes-cbc-generate-export-key-raw.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting an AES-CBC key with raw format");

jsTestIsAsync = true;

var extractable = true;

debug("Generating a key...");
crypto.subtle.generateKey({name: "aes-cbc", length: 128}, extractable, ["decrypt", "encrypt", "unwrapKey", "wrapKey"]).then(function(key) {
    debug("Exporting a key...");
    return crypto.subtle.exportKey("raw", key);
}).then(function(result) {
    key = result;
    shouldBe("key.byteLength", "16");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
